+++ chevah/compat/tests/normal/test_avatar.py
@@ -7,13 +7,14 @@ from __future__ import with_statement
 from __future__ import print_function
 from __future__ import division
 from __future__ import absolute_import
+from __future__ import unicode_literals
 
 from chevah.compat.avatar import FilesystemAvatar
 from chevah.compat.interfaces import IFileSystemAvatar
 from chevah.compat.testing import ChevahTestCase, mk
 
 
-class TestAvatarBase(ChevahTestCase):
+class TestFilesystemAvatar(ChevahTestCase):
 
     def test_init_no_arguments(self):
         """
@@ -56,20 +57,41 @@ class TestAvatarBase(ChevahTestCase):
         self.assertEqual(mk.fs.temp_path, avatar.home_folder_path)
         self.assertEqual(name, avatar.name)
         self.assertIsNone(avatar.root_folder_path)
+        self.assertEqual((), avatar.virtual_folders)
 
     def test_init_all_arguments(self):
         """
-        Avatar can also be initialized with a root path.
+        Avatar can also be initialized with argument and the exported
+        attributes are read only.
         """
         avatar = FilesystemAvatar(
             name=mk.getUniqueString(),
             home_folder_path=u'some-path',
             root_folder_path=u'other-path',
             token=u'the-token',
+            virtual_folders=(
+                (['base', 'segment'], '/some/real/path'),
+                (['other', 'segment'], 'c:\\other\\path'),
+                )
             )
 
         self.assertEqual(u'other-path', avatar.root_folder_path)
+        with self.assertRaises(AttributeError):
+            avatar.root_folder_path = 'something'
+
         self.assertEqual(u'the-token', avatar.token)
+        with self.assertRaises(AttributeError):
+            avatar.token = 'something'
+
+        self.assertEqual(
+            (
+                (['base', 'segment'], '/some/real/path'),
+                (['other', 'segment'], 'c:\\other\\path'),
+                ),
+            avatar.virtual_folders,
+            )
+        with self.assertRaises(AttributeError):
+            avatar.virtual_folders = 'something'
 
 
 class TestApplicationAvatar(ChevahTestCase):