+++ chevah/compat/unix_users.py
@@ -488,6 +488,7 @@ class UnixDefaultAvatar(UnixHasImpersonatedAvatar):
     lock_in_home_folder = False
     token = None
     peer = None
+    virtual_folders = ()
 
     @property
     def use_impersonation(self):
@@ -516,6 +517,7 @@ class UnixSuperAvatar(UnixHasImpersonatedAvatar):
     lock_in_home_folder = False
     token = None
     peer = None
+    virtual_folders = ()
 
     @property
     def use_impersonation(self):